Move allowFraming out
