+++ lib/routes/wellKnownQueries.ts
@@ -21,20 +21,18 @@ import {
 } from "@atomist/automation-client";
 import { DeliveryPhases } from "@atomist/sdm-pack-analysis/lib/analysis/phases";
 import { DockerFileParser } from "@atomist/sdm-pack-docker";
-import { CodeMetrics } from "@atomist/sdm-pack-sloc";
 import {
     CodeStats,
     consolidate,
     Language,
-    LanguageStats,
 } from "@atomist/sdm-pack-sloc/lib/slocReport";
 import { DockerStack } from "@atomist/uhura/lib/element/docker/dockerScanner";
 import * as _ from "lodash";
 import { ProjectAnalysisResult } from "../analysis/ProjectAnalysisResult";
 import { CodeMetricsElement } from "../element/codeMetricsElement";
 import { PackageLock } from "../element/packageLock";
+import { Queries } from "../feature/queries";
 import {
-    SunburstTreeEmitter,
     treeBuilder,
     TreeBuilder,
 } from "../tree/TreeBuilder";
@@ -43,7 +41,6 @@ import {
     OrgGrouper,
     ProjectAnalysisResultGrouper,
 } from "./projectAnalysisResultUtils";
-import { Queries } from "../feature/queries";
 
 /**
  * Well known queries against our repo cohort
